% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhylo.R
\name{Annoft}
\alias{Annoft}
\title{Selection and Annotation of ft matrix}
\usage{
Annoft(geneVec, inputft, allAnno)
}
\arguments{
\item{geneVec}{A vector containing condidate genes used for Circos plot. The length of one is also allowed.}

\item{inputft}{The "from-to" matrix. The 1st and 2nd columns are nodes names, and the 3rd column is the similarity. The "inputft" may contain genes not presenting the "geneVec".}

\item{allAnno}{The annotation matrix. The 1st is the KEGG gene IDs, which has the same format with "inputft". The 2nd is the symbol name shown in the Circos plot. The 3rd column is the chromosome name. The 4th and 5th columns are the start and end positions. The "allAnno" may contains genes that are absent in the "inputft".}
}
\value{
An annotated ft matrix
}
\description{
Selection and Annotation of from-to matrix
}
\details{
Annotation the "from-to matrix"(network data). The main purpose is: 1. select ft matrix containing certain gene list; 2. transfer geneID (hsa:1) to gene symbol (A1BG).
}
\examples{
data(atpft)
atpft <- atpft[, c(1, 3, 5:6)]
data(geneAnno)
f1genes <- c('hsa:498', 'hsa:506', 'hsa:509', 'hsa:539', 'hsa:513', 'hsa:514')
newft <- Annoft(f1genes, atpft, geneAnno)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
