% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDef.R
\docType{data}
\name{atpft}
\alias{atpft}
\title{Whole genome predicted linkages of human F1Fo ATP synthase subunits.}
\format{A character matrix}
\description{
A character matrix:
the 1st column and 2ed colum are gene ID;
the 3rd clumn is the Jaccard similarity ranging from 0 to 1.
}
\references{
Unpublished data from Yulong Niu
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
