% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhylo.R
\name{CheckLinkCol}
\alias{CheckLinkCol}
\title{Check linkage colors}
\usage{
CheckLinkCol(geneVec, linkColVec, allAnno)
}
\arguments{
\item{geneVec}{A vector containing condidate genes used for Circos plot. The length of one is also allowed.}

\item{linkColVec}{A vector of colors used for linkages.}

\item{allAnno}{The annotation matrix. The 1st is the KEGG gene IDs, which has the same format with "inputft". The 2nd is the symbol name shown in the Circos plot. The 3rd column is the chromosome name. The 4th and 5th columns are the start and end positions. The "allAnno" may contains genes that are absent in the "inputft".}
}
\value{
A list containing warning message.
}
\description{
Check the input genes whether or not have annotation information and whether or not have colors
}
\details{
Check the input genes with annotation or having colors. if not, an error message with length 1 will be returned.
}
\examples{
data(geneAnno)
## no annotation and containing NA
g1 <- c('hsa:111111', 'hsa:498', 'hsa:506', 'hsa:509', 'hsa:516', 'hsa:517')
g1Col <- c('blue', 'dblue_a3', 'dyellow_a3', 'dred_a3', NA, NA)
g1 <- c(NA, NA)
g1 <- c('hsa:0000', 'hsa:111111')
g1 <- c('hsa:111111', NA, NA)
g1 <- c('hsa:111111', 'hsa:498', NA, NA)
checkColList <- CheckLinkCol(g1, g1Col, geneAnno)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
