% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circosConfig.R
\name{cRules}
\alias{cRules}
\alias{cLabelGenes}
\alias{cHeatmap}
\alias{cFreq}
\alias{cLinks}
\alias{cPhyloLink}
\alias{writeConf}
\title{Generate Circos configure}
\usage{
cRules(geneSymbolVec, linkColVec)

cLabelGenes(savePath, geneSymbolVec, linkColVec)

cHeatmap(savePath, geneVec, linkColVec)

cFreq(savePath)

cLinks(savePath, geneVec, linkColVec)

cPhyloLink(savePath, geneVec, geneSymbolVec, linkColVec, organism)

writeConf(savePath, geneVec, geneSymbolVec, linkColVec, organism, storePath)
}
\arguments{
\item{geneSymbolVec}{A vector containing condidate genes symbol.}

\item{linkColVec}{A vector of colors used for linkages, names, and heatmaps.}

\item{savePath}{The folder of label files.}

\item{geneVec}{A vector containing condidate genes which should be the same as the input of function "writeCircos". CAUTION: The element in the "geneVec" will be used in the files' name!!! An example is c('hsa:516', 'hsa:517')}

\item{organism}{KEGG species ID, for example "hsa" for human.}

\item{storePath}{The store path of Circos configure file.}
}
\value{
A formated rules vector.

A formated label vector.

A formated heatmap vector.

A formated frequency

A formated links

A formated Circos configure vecor of predicted linkages.
}
\description{
Generate configures for Circos plot
}
\details{
"labelConf" generates gene labels for the Circos configures ("names and locations of genes"). "cHeatmap" generates heat map for the Circos configures ("heatmap"(correlation value) for each linkage). "cFreq" generates frequency of presence in three domains. "cLinks" generates linkages for each input "geneVec". "cPhyloLink" generates the whole file automatically.
Please NOTE this temporal configure code is a piece of shit. Please check json format and re-organize this shit.
}
\examples{
f1 <- c('ATP5A1', 'ATP5B', 'ATP5C1', 'ATP5O', 'ATP5D', 'ATP5E')
f1Col <- c('dblue_a3', 'dyellow_a3', 'dred_a3', 'dgreen_a3', 'dpurple_a3', 'dorange_a3')
labelConf <- cLabelGenes('phylo/', f1, f1Col)
f1genes <- c('hsa:498', 'hsa:506', 'hsa:509', 'hsa:539', 'hsa:513', 'hsa:514')
f1Col <- c('dblue_a3', 'dyellow_a3', 'dred_a3', 'dgreen_a3', 'dpurple_a3', 'dorange_a3')
heatConf <- cHeatmap('phylo/', f1genes, f1Col)
freqConf <- cFreq('phylo/')
f1genes <- c('hsa:498', 'hsa:506', 'hsa:509', 'hsa:539', 'hsa:513', 'hsa:514')
f1Col <- c('dblue_a3', 'dyellow_a3', 'dred_a3', 'dgreen_a3', 'dpurple_a3', 'dorange_a3')
linkConf <- cLinks('phylo/', f1genes, f1Col)
f1 <- c('ATP5A1', 'ATP5B', 'ATP5C1', 'ATP5O', 'ATP5D', 'ATP5E')
f1genes <- c('hsa:498', 'hsa:506', 'hsa:509', 'hsa:539', 'hsa:513', 'hsa:514')
f1Col <- c('dblue_a4', 'dyellow_a4', 'dred_a4', 'dgreen_a4', 'dpurple_a4', 'dorange_a4')
wholeConf <- cPhyloLink(savePath = 'phylo/', geneVec = f1,
geneSymbolVec = f1, linkColVec = f1Col, organism = 'hsa')
wholeConf <- cPhyloLink(savePath = 'phylo/', geneVec = f1genes,
geneSymbolVec = f1, linkColVec = f1Col, organism = 'hsa')

f1 <- c('ATP5A1', 'ATP5B', 'ATP5C1', 'ATP5O', 'ATP5D', 'ATP5E')
f1genes <- c('hsa:498', 'hsa:506', 'hsa:509', 'hsa:539', 'hsa:513', 'hsa:514')
f1Col <- c('dblue_a4', 'dyellow_a4', 'dred_a4', 'dgreen_a4', 'dpurple_a4', 'dorange_a4')
\dontrun{
writeConf(savePath = 'phylo/', geneVec = f1genes,
geneSymbolVec = f1, linkColVec = f1Col,
organism = 'hsa', storePath = 'tmp1/')
}
}
\seealso{
writeCircos
}
\author{
Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}
}
