% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3network.R
\name{d3ExtractNetEle}
\alias{d3ExtractNetEle}
\title{Extract d3 network}
\usage{
d3ExtractNetEle(d3htmlPath)
}
\arguments{
\item{d3htmlPath}{An absolute path of html, for example '/home/User1/Download/index.html'.}
}
\value{
A string vector containing the d3 network html elements.
}
\description{
Extract d3 network elements
}
\details{
Retrieve the d3 network elements from a html file.
}
\examples{
filePath <- system.file("extdata", "d3net.html", package = "PhyloProfileSuppl")
d3Ele <- d3ExtractNetEle(filePath)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
