% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circosGenerate.R
\name{ft2circos}
\alias{ft2circos}
\title{Generate Circos phylogenetic files}
\usage{
ft2circos(ft, locaAnno, nodeName = "all", thick = 5, showEdge = TRUE)
}
\arguments{
\item{ft}{A from-to matrix, here we used the undirected network. The 1st and 2nd columns are gene ID, and the 3rd clumn is the weight ranging from 0 to 1 like correlation coefficient or similarity value.}

\item{locaAnno}{The gene location annotation information. The 1st column is the gene ID which should be the same format in "ft". The 2nd column is the chromosome name, like "1", "2", and "MT". The 3rd and 4th columns are the start and end position, respectively.}

\item{nodeName}{Setting node names desided to show. The default value is "all" that showing all the input nodes. Users can set this parameters to a subset of the nodes.}

\item{thick}{The thickness}

\item{showEdge}{Whether or not to show the weighted thickness (wThickness)of linkages. The default value is "TRUE" that means the wThickness = thick * weight. Otherwise, the thickness of all the linkages is set the same as the input "thick".}
}
\value{
A list of Circos links, Circos lables, and Circos heatmap.
}
\description{
Generate phylogenetic data for Circos plot
}
\details{
This function is used to generate Circos links, Circos lables, and Circos heatmap.
}
\examples{
data(atpft)
data(geneAnno)
# preprocess
geneAnno <- geneAnno[, -2]
interLogic <- (atpft[, 1] \%in\% geneAnno[, 1]) & (atpft[, 2] \%in\% geneAnno[, 1])
atpft <- atpft[interLogic, ]
# show all the linkages with weighted thickness
fatpCircos <- ft2circos(ft = atpft, locaAnno = geneAnno)
# only show the links with "ATP5A1" without weighed thickness
fatpCircos <- ft2circos(ft = atpft, locaAnno = geneAnno, nodeName = 'ATP5A1', showEdge = FALSE)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
