% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDef.R
\docType{data}
\name{geneAnno}
\alias{geneAnno}
\title{Annotatio of human genes.}
\format{A character matrix}
\description{
A character matrix:
the 1st column is the gene ID which should be the same format in "atpft";
the 2nd column is the chromosome name, like "1", "2", and "MT";
the 3rd and 4th columns are the start and end position, respectively.
}
\references{
Unpublished data from Yulong Niu
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
