% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhylo.R
\name{GetLinkages}
\alias{GetLinkages}
\title{Retrieve linkages data}
\usage{
GetLinkages(geneIDs, linkData)
}
\arguments{
\item{geneIDs}{The vector of geneIDs.}

\item{linkData}{A matrix, of which the first and third columns is gene Ids.}
}
\value{
A linkage matrix.
}
\description{
Retrieve linkages
}
\details{
Retrieve the linkages containing the input gene list. This function will be replaced by a SQL query.
}
\examples{
genes <- c('a', 'c')
linkMat <- matrix(c(letters[1:5], 1:5, letters[5:1], 5:1),
ncol = 4,
nrow = 5,
dimnames = list(paste0('link', 1:5), c('From', 'FromLink', 'To', 'ToLink')))
geneLinkMat <- GetLinkages(genes, linkMat)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
