% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhylo.R
\name{GetProfile}
\alias{GetProfile}
\title{Retrieve profiles}
\usage{
GetProfile(geneIDs, profileData)
}
\arguments{
\item{geneIDs}{The vector of geneIDs.}

\item{profileData}{A numeric matrix. The row names are the genes and and the column names are the species.}
}
\value{
The selected profiles.
}
\description{
Retrieve profiles
}
\details{
Retrieve the phylogenetic profiles containing the input gene list. This function will be replaced by a SQL query.
}
\examples{
genes <- c('a', 'c')
profileMat <- matrix(sample(0:1, size = 20, replace = TRUE),
ncol = 4,
nrow = 5,
dimnames = list(letters[1:5], paste0('spe', 1:4)))
geneProfileMat <- GetProfile(genes, profileMat)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
