% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhylo.R
\name{GetTopLink}
\alias{GetTopLink}
\title{Retrieve top linkages data}
\usage{
GetTopLink(geneIDs, linkData, annoVec, threshold = 400)
}
\arguments{
\item{geneIDs}{The vector of input geneIDs.}

\item{linkData}{A list, first element is the n x m matrix, n is the top jaccard value (decreasing order), m is the whole genome genes. Second and third element is n x m matrix with Jaccard and Correlation values.}

\item{annoVec}{A vector contains m anntoations of m genes.}

\item{threshold}{A threshold should 0 < threshold <= n}
}
\value{
a formated linkages
}
\description{
Retrieve top linkages
}
\details{
Give users freedom to choose threshold of top linkages.
}
\examples{
lkName <- matrix(ncol = 10, nrow = 9)
colnames(lkName) <- paste0('g', 1:10)
for(i in 1:10) {
  rc <- sample(1:10, 10)
  rc <- rc[rc != i]
  rcn <- paste0('g', rc)
  lkName[, i] <- rcn
}
lkJac <- matrix(round(runif(9 * 10), 2), ncol = 10)
colnames(lkJac) <- colnames(lkName)
lkCor <- matrix(round(runif(9 * 10),2),  ncol = 10)
colnames(lkCor) <- colnames(lkName)
lkData <- list(lkName, lkJac, lkCor)
gnIDs <- paste0('g', c(3, 5, 7))
aoVec <- paste0('gene', 1:10)
names(aoVec) <- colnames(lkName)
canLink <- GetTopLink(gnIDs, lkData, aoVec, 4)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
