% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circosJSConfig.R
\name{LinkJS}
\alias{LinkJS}
\title{Linkages for circosJS plot}
\usage{
LinkJS(ft, allAnno)
}
\arguments{
\item{ft}{A from-to matrix, here we used the undirected network. The 1st and 2nd columns are gene ID, and other columns are similarity or distances (Jaccard, Cor et. al.).}

\item{allAnno}{The annotation matrix. The 1st is the KEGG gene IDs, which has the same format with "inputft". The 2nd is the symbol name shown in the Circos plot. The 3rd column is the chromosome name. The 4th and 5th columns are the start and end positions. The "allAnno" may contains genes that are absent in the "inputft".}
}
\value{
A linkage matrix.
}
\description{
Generate linkages for circosJS plot
}
\details{
This function is used to generate linkages for circosJS plot
}
\examples{
data(atpft)
atpft <- atpft[, c(1, 3, 5:6)]
data(geneAnno)
LinkJS(atpft, geneAnno)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
