% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3network.R
\name{Normalize}
\alias{Normalize}
\title{normalization}
\usage{
Normalize(inputVec, addSmall = 0)
}
\arguments{
\item{inputVec}{The input raw vector.}

\item{addSmall}{A small value used to add on the normalized data.}
}
\value{
normalized vector
}
\description{
Normalization of the numeric vector
}
\details{
The normalization formular is (x - min(X))/(max(X) - min(X)). "NA"s are removed. If the max(x) is equal to min(x), asign all the elements to 0.5.
}
\examples{
testVec <- sample(1:100, 10)
Normalize(testVec)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
