% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDef.R
\docType{data}
\name{phyloSpe}
\alias{phyloSpe}
\title{Phylogeny of 972 KEGG species}
\format{A character matrix}
\description{
A character matrix:
the 1st column is KEGG species IDs;
the 2nd column is the corresponding phylogeny;
}
\references{
Unpublished data from Yulong Niu
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
