% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3network.R
\name{d3Transft}
\alias{d3Transft}
\alias{d3PlotNet}
\alias{writed3Net}
\title{plot D3 network}
\usage{
d3Transft(d3Annoft)

d3PlotNet(d3ft, linkStrengthBase = 5, ...)

writed3Net(d3NetObj, fileName = "networkd3.html", savePath = "./")
}
\arguments{
\item{d3Annoft}{A ft matrix used for d3 plot. 1st column: from nodes. 2nd column: to nodes. 3rd column: linkage strength, for example the Jaccard similarity ranging from 0 to 1. 4th column: linkage colours.

A normalized linkage strength is used. The nodes are grouped by different color, and the latter gene will cover the former one. The size of a node is determined by the number of its partners.}

\item{d3ft}{The d3 ft list from d3Transft().}

\item{linkStrengthBase}{The base value of linkage strength.}

\item{...}{Additional paramters from forceNetwork().}

\item{d3NetObj}{A d3 network object from PlotNetworkD3().}

\item{fileName}{The file name to store the html.}

\item{savePath}{The save path of html.}
}
\value{
d3Transft returns a list, the first element is links dataframe and the second element is the nodes dataframe.
PlotNetworkD3 returns an object of D3 network.
writeD3network returns nothing.
}
\description{
Plot network using the javascript D3 library
}
\details{
D3 library is used for network plot.
}
\examples{
filePath <- system.file("extdata", "linkage.csv", package = "pppSupplR")
atpft <- read.csv(filePath, stringsAsFactor = FALSE)
atpft <- atpft[, c(4, 8, 9, 11)]
newd3ft <- d3Transft(atpft)
d3Obj <- d3PlotNet(newd3ft)
\dontrun{
writed3Net(d3Obj, 'testd3.html')
}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
