% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circosGenerate.R
\name{SpeFreq}
\alias{SpeFreq}
\title{Generate gene frequency}
\usage{
SpeFreq(KEGGPhylo, selectPhyloData, splitEu = FALSE)
}
\arguments{
\item{KEGGPhylo}{A character matrix, the 1st column is the KEGGID, and the 2nd column is species phylogeny for example "Eukaryotes;Animals;Vertebrates;Mammals".}

\item{selectPhyloData}{A numeric matrix. The row names are the genes and and the column names are the species.}

\item{splitEu}{Whether to show the eukaryotic split, the default value is "FALSE".}
}
\value{
A fequency matrix.
}
\description{
Generate gene frequency from phylogenetic profiles.
}
\details{
Calculate gene frequence of three domains of life or more details of Eukaryota.
}
\examples{
data(phyloSpe)
data(wholeProfile)
freqMat <- SpeFreq(KEGGPhylo = phyloSpe, selectPhyloData = wholeProfile, splitEu = TRUE)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
