% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circosJSConfig.R
\name{SpeFreqJS}
\alias{SpeFreqJS}
\title{Generate gene frequency for circosJS}
\usage{
SpeFreqJS(allSpePhylo, allPhyloData, allAnno, splitEu = FALSE)
}
\arguments{
\item{allSpePhylo}{The species and phylogeny matrix.}

\item{allPhyloData}{The phylogenetic profiles.}

\item{allAnno}{The annotation matrix. The 1st is the KEGG gene IDs, which has the same format with "inputft". The 2nd is the symbol name shown in the Circos plot. The 3rd column is the chromosome name. The 4th and 5th columns are the start and end positions. The "allAnno" may contains genes that are absent in the "inputft".}

\item{splitEu}{Whether to show the eukaryotic split, the default value is "FALSE".}
}
\value{
A fequency matrix.
}
\description{
Generate gene frequency for circosJS from phylogenetic profiles.
}
\details{
Calculate gene frequence of three domains of life or more details of Eukaryota.
}
\examples{
data(phyloSpe)
data(wholeProfile)
data(geneAnno)
freqMat <- SpeFreqJS(phyloSpe, wholeProfile, geneAnno, splitEu = TRUE)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
