% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDef.R
\docType{data}
\name{wholeProfile}
\alias{wholeProfile}
\title{Phylogenic profiles}
\format{A character matrix}
\description{
A character matrix:
the column names are the KEGG species IDs;
the row names are the KEGG gene IDs;
}
\references{
Unpublished data from Yulong Niu
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
