% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_writeCirosFile.R
\name{writeCircos}
\alias{writeCircos}
\title{Write circos data files}
\usage{
writeCircos(geneVec, inputft, allAnno, allSpePhylo, allPhyloData, savePath)
}
\arguments{
\item{geneVec}{A vector containing condidate genes used for Circos plot. The length of one is also allowed.}

\item{inputft}{The "from-to" matrix. The 1st and 2nd columns are nodes names, and the 3rd column is the similarity. The "inputft" may contain genes not presenting the "geneVec".}

\item{allAnno}{The annotation matrix. The 1st is the KEGG gene IDs, which has the same format with "inputft". The 2nd is the symbol name shown in the Circos plot. The 3rd column is the chromosome name. The 4th and 5th columns are the start and end positions. The "allAnno" may contains genes that are absent in the "inputft".}

\item{allSpePhylo}{The species and phylogeny matrix.}

\item{allPhyloData}{The phylogenetic profiles.}

\item{savePath}{The path used to store Circos files.}
}
\description{
Automatically write circos data files for plotting linkages
}
\details{
Generate four types circos data files:
1. linkages for each input "geneVec";
2. "heatmap"(correlation value) for each linkage;
3. frequency of presence in three domains;
4. names and locations of genes
}
\examples{
\dontrun{
data(atpft)
data(geneAnno)
data(phyloSpe)
data(wholeProfile)
f1genes <- c('hsa:498', 'hsa:506', 'hsa:509', 'hsa:539', 'hsa:513', 'hsa:514')
f0genes <- c('hsa:516', 'hsa:517', 'hsa:518', 'hsa:515', 'hsa:521',
'hsa:522', 'hsa:9551', 'hsa:10476', 'hsa:10632', 'hsa:4508', 'hsa:4509')
fgenes <- c(f1genes, f0genes)
tmp1 <- writeCircos(f1genes, atpft, geneAnno, phyloSpe, wholeProfile, savePath = 'tmp1/')}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
