% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circosJSConfig.R
\name{writeCircosJS}
\alias{writeCircosJS}
\title{Write circosJS data files}
\usage{
writeCircosJS(geneVecList, ft, allAnno, allSpePhylo, allPhyloData, savePath)
}
\arguments{
\item{geneVecList}{The output object from the "CheckLinkCol" function.}

\item{ft}{A from-to matrix, here we used the undirected network. The 1st and 2nd columns are gene ID, and other columns are similarity or distances (Jaccard, Cor et. al.).}

\item{allAnno}{The annotation matrix. The 1st is the KEGG gene IDs, which has the same format with "inputft". The 2nd is the symbol name shown in the Circos plot. The 3rd column is the chromosome name. The 4th and 5th columns are the start and end positions. The "allAnno" may contains genes that are absent in the "inputft".}

\item{allSpePhylo}{The species and phylogeny matrix.}

\item{allPhyloData}{The phylogenetic profiles.}

\item{savePath}{The path used to store circosJS files.}
}
\description{
Automatically write circosJS data files for plotting linkages
}
\details{
Generate two types circosJS data files:
1. linkages for each input "geneVec";
2. frequency of presence in three domains;
}
\examples{
\dontrun{
data(atpft)
atpft <- atpft[, c(1, 3, 5:6)]
data(geneAnno)
data(phyloSpe)
data(wholeProfile)
g1 <- c('hsa:111111', 'hsa:498', 'hsa:506', 'hsa:509', 'hsa:516', 'hsa:517')
g1Col <- c('#36A03F', '#49AE4F', '#1C386A', '#A03531', NA, NA)
checkList <- CheckLinkCol(g1, g1Col, geneAnno)
tmp1 <- writeCircosJS(checkList, atpft, geneAnno, phyloSpe, wholeProfile, savePath = 'tmp1/')}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
